requirements.txt
selenium
beautifulsoup4
elasticsearch
pandas
webdriver_manager
fake_useragent
Pillow
requests
imagehash
scikit-learn
pyppeteer
numpy
fake_useragent
html2text
python-dotenv
shopifyapi